requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anthropic==0.2.10
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
exceptiongroup==1.1.3
filelock==3.13.0
frozenlist==1.4.0
fsspec==2023.10.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
multidict==6.0.4
openai==0.27.10
packaging==23.2
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
tenacity==8.2.3
tokenizers==0.14.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2